code_RobustLoc/requirements.txt
colour_demosaicing==0.2.2
matplotlib==3.3.4
numpy==1.19.2
Pillow==9.2.0
scipy==1.5.4
torch==1.10.0
torchdiffeq==0.2.3
torchvision==0.11.1
tqdm==4.64.0
transforms3d==0.3.1
